Fixed a bug in bin step type conversion.

+++ svl/svl.py
@@ -74,7 +74,7 @@ class SVLToVegaLiteTransformer(lark.Transformer):
         if len(items) == 0:
             return {"bin": True}
         else:
-            return {"bin": {"step": int(items[0])}}
+            return {"bin": {"step": float(items[0])}}
 
     def aggregation(self, items):
         return {
